requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.17.0
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appdirs==1.4.4
asttokens==3.0.0
astunparse==1.6.3
asyncssh==2.21.1
atpublic==6.0.2
attrs==25.4.0
Authlib==1.6.5
backoff==2.2.1
billiard==4.2.2
blinker==1.9.0
blis==0.7.11
boto3==1.40.53
botocore==1.40.53
cachetools==6.2.1
catalogue==2.0.10
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.16.0
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
confection==0.1.5
configobj==5.0.9
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
cyclopts==3.24.0
cymem==2.0.11
dacite==1.6.0
dagshub==0.6.3
dagshub-annotation-converter==0.1.13
databricks-sdk==0.68.0
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.2
dpath==2.2.0
dulwich==0.24.5
dvc==3.63.0
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.2
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
email-validator==2.3.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
faiss-cpu==1.12.0
fastapi==0.119.0
fastjsonschema==2.21.2
fastmcp==2.12.4
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
funcy==2.0
gast==0.6.0
gensim==4.3.2
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
google-pasta==0.2.0
gql==4.0.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
grpcio==1.75.1
gto==1.9.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.0
kagglehub==0.3.13
keras==3.11.3
kiwisolver==1.4.9
kombu==5.5.4
langchain==0.3.27
langchain-community==0.3.16
langchain-core==0.3.79
langchain-text-splitters==0.3.11
langcodes==3.5.0
langsmith==0.3.45
language_data==1.3.0
lazy-object-proxy==1.12.0
libclang==18.1.1
lxml==6.0.2
Mako==1.3.10
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.8.2
matplotlib-inline==0.1.7
mcp==1.17.0
mdurl==0.1.2
ml-dtypes==0.4.1
mlflow==3.5.0
mlflow-skinny==3.5.0
mlflow-tracing==3.5.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
murmurhash==1.0.13
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.8.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.8.1
numpy==1.25.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.4.107
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.2.0.103
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opt_einsum==3.4.0
optree==0.17.0
optuna==4.5.0
optuna-integration==4.6.0
orjson==3.11.3
packaging==25.0
pandas==2.1.4
parse==1.20.2
parso==0.8.5
pathable==0.4.4
pathspec==0.12.1
pathvalidate==3.3.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
plotly==5.17.0
preshed==3.0.10
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==4.25.8
psutil==7.1.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
pydot==4.0.1
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
pymongo==4.15.3
pyparsing==3.2.5
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==6.4.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
s3transfer==0.14.0
scikit-learn==1.3.2
scipy==1.11.4
scmrepo==3.5.2
seaborn==0.13.0
semver==3.0.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
sqlparse==0.5.3
sqltrie==0.11.2
srsly==2.5.1
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
streamlit==1.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
textblob==0.17.1
thinc==8.2.5
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.3
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
treelib==1.8.0
triton==3.3.1
typer==0.9.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
vine==5.1.0
voluptuous==0.15.2
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.14
weasel==0.3.4
Werkzeug==3.1.1
wordcloud==1.9.2
wrapt==1.17.3
xlsxwriter==3.2.9
yarl==1.22.0
zc.lockfile==4.0
zipp==3.23.0
zstandard==0.23.0
